# openapi-example-1.0.0.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/products": {
            "get": {
                "operationId": "ProductController_getProducts",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "products"
                ]
            },
            "post": {
                "operationId": "ProductController_addProduct",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductId"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductId"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "products"
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "operationId": "ProductController_getProduct",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "products"
                ]
            },
            "patch": {
                "operationId": "ProductController_patchProduct",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "products"
                ]
            },
            "delete": {
                "operationId": "ProductController_deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductId"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "products"
                ]
            }
        },
...
    },
    "info": {
        "title": "Product API",
        "description": "An API for products and learn NestJS",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "itemVariants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ItemVariant"
                        }
                    }
                },
                "required": [
                    "id",
                    "title",
                    "description",
                    "price",
                    "itemVariants"
                ]
            },
            "ItemVariant": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "size": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    }
                },
                "required": [
                    "id",
                    "size",
                    "color",
                    "product"
                ]
            },
            "ItemVariantData": {
                "type": "object",
                "properties": {
                    "size": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    }
                },
                "required": [
                    "size",
                    "color"
                ]
            },
            "ProductData": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "itemVariants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ItemVariantData"
                        }
                    }
                },
                "required": [
                    "title",
                    "description",
                    "price",
                    "itemVariants"
                ]
            },
            "ProductId": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    }
                },
                "required": [
                    "id"
                ]
            },
...
        }
    }
}